requirements.txt
openai>=1.66.3
python-dotenv>=1.0.1
